Add dynamic module build for Debian Stretch and Ubuntu Zesty.
This also removes the pkg:nginx target since the dyn module works with
the distro's Nginx, no need for our own package. 

For Stretch, we ensure that we link the Passenger pkg to the same
openssl version as libcurl.

All passenger source packages are now labeled by distro, and all package
descriptions were updated and improved.